// packages/primitives/avatar/src/types.ts
export type RdxImageLoadingStatus = 'idle' | 'loading' | 'loaded' | 'error';

export type HTMLAttributeReferrerPolicy =
    | ''
    | 'no-referrer'
    | 'no-referrer-when-downgrade'
    | 'origin'
    | 'origin-when-cross-origin'
    | 'same-origin'
    | 'strict-origin'
    | 'strict-origin-when-cross-origin'
    | 'unsafe-url';